requirements.txt
# Core dependencies
requests==2.28.2
pygithub==1.58.1
python-dotenv==1.0.0
click==8.1.3

# Code analysis tools
pylint==2.17.0
flake8==6.0.0
bandit==1.7.5
mypy==1.1.1
isort==5.12.0
black==23.1.0

# Testing
pytest==7.3.1

# Documentation
sphinx==6.1.3